+++ app/model/projects.js
@@ -194,7 +194,7 @@ var Projects = {
                 "       s.token_created_on \n" +
                 "FROM sites AS s \n"+
                 "LEFT JOIN project_imported_sites as pis ON s.site_id = pis.site_id AND pis.project_id = ? \n"+
-                "WHERE (s.project_id = ? OR pis.project_id = ?)",  typeCast: sqlutil.parseUtcDatetime },
+                "WHERE (s.project_id = ? OR pis.project_id = ?) AND s.deleted_at is null",  typeCast: sqlutil.parseUtcDatetime },
                 [project_id, project_id, project_id, project_id]
         ).then(function(sites){
             if(sites.length && options && options.compute){
@@ -245,7 +245,7 @@ var Projects = {
             FROM sites AS s
             LEFT JOIN project_imported_sites as pis ON s.site_id = pis.site_id AND pis.project_id=${project_id}
             LEFT JOIN recordings as r ON s.site_id = r.site_id
-            WHERE (s.project_id=${project_id} OR pis.project_id=${project_id})
+            WHERE (s.project_id=${project_id} OR pis.project_id=${project_id}) AND s.deleted_at is null
             GROUP BY s.site_id ORDER by s.name;`;
         return dbpool.streamQuery({ sql })
     },
@@ -254,7 +254,7 @@ var Projects = {
         return dbpool.query({ sql:
             `SELECT COALESCE(DATE_FORMAT(r.datetime, "${format}")) as date
             FROM recordings AS r
-            JOIN sites as s ON s.site_id = r.site_id AND s.project_id = ?
+            JOIN sites as s ON s.site_id = r.site_id AND s.project_id = ? AND s.deleted_at is null
             LEFT JOIN project_imported_sites as pis ON s.site_id = pis.site_id AND pis.project_id = ?
             WHERE s.project_id = ? OR pis.project_id = ?
             GROUP BY YEAR(r.datetime), MONTH(r.datetime), DAY(r.datetime) ORDER BY r.datetime ASC`,  typeCast: sqlutil.parseUtcDatetime },
@@ -996,7 +996,7 @@ var Projects = {
     // this includes recordings processing
     totalRecordings: function(project_id) {
         var q = "SELECT count(recording_id) as count \n" +
-                "FROM recordings AS r JOIN sites AS s ON s.site_id = r.site_id \n"+
+                "FROM recordings AS r JOIN sites AS s ON s.site_id = r.site_id AND s.deleted_at is null \n"+
                 "WHERE s.project_id = " + dbpool.escape(project_id);
 
         return dbpool.query(q).get(0).get('count');