+++ scripts/db/033-add-deleted-at-to-sites.sql
@@ -1,0 +1,4 @@
+ALTER TABLE `arbimon2`.`sites`
+ADD COLUMN `deleted_at` DATETIME DEFAULT NULL;
+
+CREATE INDEX deleted_at ON sites(deleted_at);